improve action code & add some more logging
